pyproject.toml
[project]
name = "mcp-mock-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckduckgo-mcp-server>=0.1.1",
    "fast-agent-mcp>=0.2.25",
    "fastapi>=0.123.0",
    "fastmcp>=2.13.1",
    "google-genai>=1.52.0",
    "oracledb>=3.4.1",
    "pymilvus>=2.6.4",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "uvicorn[standard]>=0.38.0",
]
